shuffle=None

+++ fiw_dataset.py
@@ -94,7 +94,7 @@ class FaceDataSet(Dataset):
 
         should_same = index % 2 == 0
         if should_same:
-            p1, p2 = self.relations[index / 2]
+            p1, p2 = self.relations[int(index / 2)]
             img1 = loader(choice(self.label_images_map[p1]), self.kind, self.argument)
             img2 = loader(choice(self.label_images_map[p2]), self.kind, self.argument)
             return img1, img2, torch.Tensor([1])
